+++ js/api.js
@@ -3,11 +3,6 @@ import {
   showAlert,
 } from './util.js';
 
-import {
-  createSuccesMessageUpload,
-  createErrorMessageUpload,
-} from './alerts-render.js';
-
 const getData = (onSuccess) => {
   fetch('https://27.javascript.pages.academy/kekstagram-simple/data')
     .then((response) => {
@@ -35,15 +30,12 @@ const sendData = (onSuccess, onFail, body) => {
     .then((response) => {
       if (response.ok) {
         onSuccess();
-        createSuccesMessageUpload();
       } else {
         onFail();
-        createErrorMessageUpload();
       }
     })
     .catch(() => {
       onFail();
-      createErrorMessageUpload();
     });
 };
